split const and logger from main
