+++ pialert/const.py
@@ -1,0 +1,21 @@
+""" CONSTANTS for Pi.Alert """
+
+#===============================================================================
+# PATHS
+#===============================================================================
+# pialertPath = '/home/pi/pialert'
+pialertPath ='/home/roland/repos/Pi.Alert'
+
+confPath = "/config/pialert.conf"
+dbPath = '/db/pialert.db'
+
+
+pluginsPath  = pialertPath + '/front/plugins'
+logPath      = pialertPath + '/front/log'
+fullConfPath = pialertPath + confPath
+fullDbPath   = pialertPath + dbPath
+fullPholusPath = pialertPath+'/pholus/pholus3.py'
+
+vendorsDB              = '/usr/share/arp-scan/ieee-oui.txt'
+piholeDB               = '/etc/pihole/pihole-FTL.db'
+piholeDhcpleases       = '/etc/pihole/dhcp.leases'
